Restructure stages to match MEDS-Transforms convention (drop explicit examples_dir)
Per mmcdermott's review on #86, reorganize tokenization and tensorization into
per-stage subpackages mirroring the upstream layout
(`MEDS_transforms/stages/<stage>/<stage>.py` + `<stage>/examples/`):

    preprocessing/
      tokenization/
        __init__.py          # re-exports `main`
        tokenization.py      # the @Stage.register'd function
        examples/
          __init__.py
          in.yaml
          out_data.yaml
      tensorization/         # same shape
        ...

Benefits:

- `Stage.register` now **auto-infers** `examples_dir` (the "file lives in a
  directory of the same name as the stage and has an `examples/` subdir" rule).
  Both decorators drop to `@Stage.register(is_metadata=False,
  example_class=MTDStageExample)` — one line each, no path wrangling.
- MEDS-Transforms' automatic docgen should pick our stages up the same way
  it picks up upstream's.
- Import path `meds_torchdata.preprocessing.tokenization` / `...tensorization`
  stays stable for callers: each subpackage's `__init__.py` re-exports `main`
  (matching upstream's `from .reshard_to_split import main as stage` pattern),
  so the `[project.entry-points."MEDS_transforms.stages"]` entries don't move.

Also drops `[tool.setuptools.package-data]` for the example YAMLs — upstream
doesn't declare it either (setuptools-scm auto-includes git-tracked files),
and my earlier explicit config was redundant with that. Verified the YAMLs
still ship in the wheel via `uv build --wheel` + `unzip -l`.
